Made collaborative ranking and refactor code. Still need to get the user names.
