+++ ac/ac-ranking/src/ActivityRunner/Justification.jsx
@@ -1,0 +1,26 @@
+// @flow
+
+import React from 'react';
+import { ReactiveText, type ActivityRunnerT } from 'frog-utils';
+
+export default (props: ActivityRunnerT) => {
+  const { activityData: { config }, logger, dataFn } = props;
+  if (!config.justify) {
+    return null;
+  } else {
+    return (
+      <ReactiveText
+        type="textarea"
+        path={['justification']}
+        logger={logger}
+        dataFn={dataFn}
+        placeholder="Justify your answer"
+        style={{
+          fontSize: '18px',
+          width: '500px',
+          height: '100px'
+        }}
+      />
+    );
+  }
+};
